file i/o

+++ 0x15-file_io/1-create_file.c
@@ -12,7 +12,7 @@
 int create_file(const char *filename, char *text_content)
 {
 	int fd;
-	ssize_t bytes_written;
+	int bytes_written;
 
 	if (filename == NULL)
 	{
